// src/api/GIPHY/schema.ts
export interface Root {
  data: GIFObject[];
  pagination: Pagination;
  meta: Meta;
}

export interface GIFObject {
  type: string;
  id: string;
  slug: string;
  url: string;
  bitly_gif_url: string;
  bitly_url: string;
  embed_url: string;
  username: string;
  source: string;
  rating: string;
  content_url: string;
  source_tld: string;
  source_post_url: string;
  is_sticker: number;
  import_datetime: string;
  trending_datetime: string;
  images: Images;
  title: string;
  analytics: Analytics;
}

export interface Images {
  fixed_height_still: FixedHeightStill;
  original_still: OriginalStill;
  fixed_width: FixedWidth;
  fixed_height_small_still: FixedHeightSmallStill;
  fixed_height_downsampled: FixedHeightDownsampled;
  preview: Preview;
  fixed_height_small: FixedHeightSmall;
  downsized_still: DownsizedStill;
  downsized: Downsized;
  downsized_large: DownsizedLarge;
  fixed_width_small_still: FixedWidthSmallStill;
  preview_webp: PreviewWebp;
  fixed_width_still: FixedWidthStill;
  fixed_width_small: FixedWidthSmall;
  downsized_small: DownsizedSmall;
  fixed_width_downsampled: FixedWidthDownsampled;
  downsized_medium: DownsizedMedium;
  original: Original;
  fixed_height: FixedHeight;
  looping: Looping;
  original_mp4: OriginalMp4;
  preview_gif: PreviewGif;
  '480w_still': N480WStill;
}
// ...
export interface Analytics {
  onload: Onload;
  onclick: Onclick;
  onsent: Onsent;
}

export interface Onload {
  url: string;
}

export interface Onclick {
  url: string;
}

export interface Onsent {
  url: string;
}

export interface Pagination {
  total_count: number;
  count: number;
  offset: number;
}

export interface Meta {
  status: number;
  msg: string;
  response_id: string;
}